backend/requirements.txt
# Django
Django==4.2.10  # Version LTS 

# Django REST Framework
djangorestframework==3.14.0

# PostgreSQL adapter
psycopg2-binary==2.9.9

# WSGI server for production
gunicorn==21.2.0

# Security and performance (recommandé)
django-cors-headers==4.3.1
whitenoise==6.6.0

# Environment variables
python-dotenv==1.0.0

dj-database-url==2.0.0